Add function to return possible options
In computer mode, the computer will take the guess, then turn the options available up to that point into an array to be returned for next evaluation. The computer calculates the possible returns for the guess and matches that to the actual returned feedback. Those options matching the feedback will be kept and fed back into the loop.