Fix method for inputting colors
Turns out that when you didn't type one of the colors in the list, it would ask you to type it again, but would also save some of the mistakes. This is now fixed in a seperate method, that only returns the color for that position once there is a valid input.